v11
v11

+++ My project/Assets/Scripts/Vida.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public class Vida : MonoBehaviour
 {
     public bool luz = false;
-    public double recuperacaoDeVida = 3;
-    public double dano = 1;
-    public double vida = 100;
+    public float recuperacaoDeVida = 3;
+    public float dano = 1;
+    public float vida = 100;
 
     // Start is called before the first frame update
     void Start()
